refactor: split DetectorMutationTests into one file per nested suite
